split out logic
